Restructure, moved Colour into structures, added MapData
