+++ src/simulator/views/map_displays/solid_color_map_display.py
@@ -6,8 +6,7 @@ from algorithms.configuration.entities.entity import Entity
 from algorithms.configuration.maps.map import Map
 from simulator.services.services import Services
 from simulator.views.map_displays.map_display import MapDisplay
-from structures import Point
-from simulator.views.map.colour import Colour
+from structures import Point, Colour
 
 
 class SolidColorMapDisplay(MapDisplay):
